requirements.txt
# CRUZ AI Complete Dependencies

# Core Web Frameworks
Flask>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Line Bot
line-bot-sdk>=3.5.0

# AI & ML
google-generativeai>=0.3.2
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# API & Data
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
websockets>=12.0
httpx>=0.25.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Discord Bot
discord.py>=2.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
click>=8.1.0
rich>=13.7.0
python-dateutil>=2.8.0fastapi
uvicorn[standard]
psycopg2-binary
asyncpg
notion-client
